+++ docs/_static/templates/degradation_api.py
@@ -1,6 +1,6 @@
 """..."""
 
-import perun.check as check
+import perun.check.factory as check
 
 
 def my_degradation_checker(baseline_profile, target_profile):